added authentication

+++ EduNote.API/Controllers/SectionsController.cs
@@ -41,7 +41,7 @@ namespace EduNote.API.Controllers
         }
 
         [HttpGet("{id}")]
-        public override IActionResult Get(int id)
+        public override IActionResult Get(long id)
         {
             try
             {
